rename generic database functions
